// src/components/vega-table/enums.ts
export enum ColumnTypes {
  Id = 'Id',
  Splitter = 'Splitter',
  Number = 'Number',
  NotRemovableWithEditor = 'NotRemovableWithEditor',
  RemovableWithEditor = 'RemovableWithEditor',
  WithClickEditor = 'WithClickEditor',
  None = 'None',
};

export enum RowTypes {
  Default = 'Default',
  Delimeter = 'Delimeter',
};

export enum SortDirection {
  ASC = 'ASC',
  DESC = 'DESC',
  UNSET = 'UNSET',
}

export enum VisibleKeys {
  Table = 'table',
}